fix: arruma validacao para a service

+++ src/domain/core/src/movie/service/DeleteMovie.ts
@@ -7,6 +7,11 @@ export default class DeleteMovie implements UseCase<string, void | null> {
 
   async execute(id: string): Promise<void | null> {
     if (!Id.isValid(id)) return null;
+    const movieExists = await this.movieRepository.findMyId(id);
+
+    if (!movieExists) {
+      throw new Error("Movie not found");
+    }
 
     return this.movieRepository.delete(id);
   }
